movies_service/requirements.txt
cassandra-driver
flask
requests